Update views.py
Fix for TeamMode where the quad is fully colored. However i cant seem to check if its for TeamMode or TA, because Team_Id is on connect -1 in TA but in Teams its 1 or 0.

However when going to teammode it seems to be correct but when the gamemode changes it doesn't return properly. (Think a script bug or a dedicated bug)