Fix view sort crash on None values
If sorting None values, the view crash with the following error :
TypeError: '<' not supported between instances of 'NoneType' and 'NoneType'
This is the case when sorting by Rank, Local, or Diff in the Map view, tab Advanced, as they can be empty on some maps.
Fix this by using float('inf') to the sort lambda if the index value is None. Effectively pushing the entry at the back of the sort.